move override logic into ld lib wrapper, and add a mock to the component tests
