+++ src/components/FeatureFlagRenderer.js
@@ -1,10 +1,8 @@
 /* @flow */
 import React, { Component } from "react";
 
-import { ldBrowserInit } from "../lib/launchDarkly";
+import { ldBrowserInit, ldOverrideFlag } from "../lib/launchDarkly";
 import { FeatureFlagType } from "../types/FeatureFlag";
-
-const url = require('url');
 
 type LaunchDarklyConfig = {
   clientId: String,
@@ -63,22 +61,7 @@ export default class FeatureFlagRenderer extends Component {
     ldClient.on("ready", () => {
       const showFeature = ldClient.variation(flagKey, false);
       const defaultState = { checkFeatureFlagComplete: true };
-
-      let override;
-      /**
-       * Follow this overriding convention:
-       * @link https://git.corp.tc/capsela/tc_feature_flagging#overriding-feature-flags
-       **/
-      const query = url.parse(window.location.toString(), true).query;
-      const queryFlag = query["features." + flagKey];
-
-      if (typeof queryFlag !== "undefined"){
-        if (queryFlag === ""){
-          override = true;
-        } else if (queryFlag === "false"){
-          override = false;
-        }
-      }
+      const override = ldOverrideFlag(flagKey);
 
       if (typeof override !== "undefined"){
         this.setState({ showFeature: override, ...defaultState});
