+++ test/components/FeatureFlagRenderer.test.js
@@ -151,6 +151,7 @@ describe("components/FeatureFlagRenderer", () => {
   describe("the _checkFeatureFlag function", () => {
 
     const variation = stub();
+    const override = stub().returns(void 0);
     const getWrapper = () => {
       return mount(
         <FeatureFlagRenderer
@@ -170,6 +171,7 @@ describe("components/FeatureFlagRenderer", () => {
         variation: variation
       });
       stub(launchDarkly, "ldBrowserInit", ldClientStub);
+      stub(launchDarkly, "ldOverrideFlag", override);
     });
 
     context("when the feature flag comes back true", () => {
@@ -188,16 +190,18 @@ describe("components/FeatureFlagRenderer", () => {
       });
     });
 
-    describe("query param flag overrides", () => {
+    describe("query param flag overrides if not undefined", () => {
       it("param \"features.flag=false\" overrides LD data \"on\"", () => {
         variation.returns(true);
-        window.location.assign("http://ab.cdef.com?features.my-test=false");
+        override.returns(false);
+        //window.location.assign("http://ab.cdef.com?features.my-test=false");
         const wrapper = getWrapper();
         expect(wrapper.state()).to.deep.equal({ checkFeatureFlagComplete: true, showFeature: false });
       });
       it("param \"features.flag\" overrides LD data \"off\"", () => {
         variation.returns(false);
-        window.location.assign("http://ab.cdef.com?features.my-test");
+        override.returns(true);
+        //window.location.assign("http://ab.cdef.com?features.my-test");
         const wrapper = getWrapper();
         expect(wrapper.state()).to.deep.equal({ checkFeatureFlagComplete: true, showFeature: true });
       });
